+++ src/modules/proto_dhcpv4/ncc_util.h
@@ -169,7 +169,7 @@ float ncc_timeval_to_float(struct timeval *in);
 int ncc_float_to_timeval(struct timeval *out, float in);
 double ncc_fr_time_to_float(fr_time_delta_t in);
 fr_time_t ncc_float_to_fr_time(double in);
-bool ncc_str_to_float(float *out, char const *in, bool allow_negative);
+bool ncc_str_to_float(double *out, char const *in, bool allow_negative);
 bool ncc_str_to_uint32(uint32_t *out, char const *in);
 size_t ncc_str_trim(char *out, char const *in, size_t inlen);
